Simple Threads Example
The first is the main thread that every Java application has. The main
thread creates a new thread from the Runnable object, MessageLoop, and
waits for it to finish. If the MessageLoop thread takes too long to
finish, the main thread interrupts it.

The MessageLoop thread prints out a series of messages. If interrupted
before it has printed all its messages, the MessageLoop thread prints a
message and exits.